Refactor FormKit input prop types and update dependencies
