+++ ui/src/formkit/inputs/attachment-group-select.ts
@@ -1,4 +1,5 @@
 import type { FormKitNode, FormKitTypeDefinition } from "@formkit/core";
+import type { FormKitInputs } from "@formkit/inputs";
 import { coreApiClient } from "@halo-dev/api-client";
 import { select } from "./select";
 
@@ -25,3 +26,12 @@ export const attachmentGroupSelect: FormKitTypeDefinition = {
   forceTypeProp: "select",
   features: [optionsHandler],
 };
+
+declare module "@formkit/inputs" {
+  export interface FormKitInputProps<Props extends FormKitInputs<Props>> {
+    attachmentGroupSelect: {
+      type: "attachmentGroupSelect";
+      value?: string;
+    };
+  }
+}
