+++ ui/src/formkit/inputs/attachment-policy-select.ts
@@ -1,4 +1,5 @@
 import type { FormKitNode, FormKitTypeDefinition } from "@formkit/core";
+import type { FormKitInputs } from "@formkit/inputs";
 import { coreApiClient } from "@halo-dev/api-client";
 import { select } from "./select";
 
@@ -22,3 +23,12 @@ export const attachmentPolicySelect: FormKitTypeDefinition = {
   forceTypeProp: "select",
   features: [optionsHandler],
 };
+
+declare module "@formkit/inputs" {
+  export interface FormKitInputProps<Props extends FormKitInputs<Props>> {
+    attachmentPolicySelect: {
+      type: "attachmentPolicySelect";
+      value?: string;
+    };
+  }
+}
