+++ ui/src/formkit/inputs/attachment/index.ts
@@ -1,4 +1,4 @@
-import { initialValue } from "@formkit/inputs";
+import { initialValue, type FormKitInputs } from "@formkit/inputs";
 import { createInput } from "@formkit/vue";
 import AttachmentInput from "./AttachmentInput.vue";
 
@@ -8,3 +8,12 @@ export const attachment = createInput(AttachmentInput, {
   forceTypeProp: "text",
   features: [initialValue],
 });
+
+declare module "@formkit/inputs" {
+  export interface FormKitInputProps<Props extends FormKitInputs<Props>> {
+    attachment: {
+      type: "attachment";
+      value?: string;
+    };
+  }
+}
