+++ ui/src/formkit/inputs/code/index.ts
@@ -1,3 +1,4 @@
+import type { FormKitInputs } from "@formkit/inputs";
 import { createInput } from "@formkit/vue";
 import CodeInput from "./CodeInput.vue";
 
@@ -6,3 +7,12 @@ export const code = createInput(CodeInput, {
   props: ["height", "language"],
   forceTypeProp: "textarea",
 });
+
+declare module "@formkit/inputs" {
+  export interface FormKitInputProps<Props extends FormKitInputs<Props>> {
+    code: {
+      type: "code";
+      value?: string;
+    };
+  }
+}
