+++ ui/src/formkit/inputs/singlePage-select.ts
@@ -1,5 +1,6 @@
 import { singlePageLabels } from "@/constants/labels";
 import type { FormKitNode, FormKitTypeDefinition } from "@formkit/core";
+import type { FormKitInputs } from "@formkit/inputs";
 import { consoleApiClient } from "@halo-dev/api-client";
 import { select } from "./select";
 
@@ -63,3 +64,12 @@ export const singlePageSelect: FormKitTypeDefinition = {
   forceTypeProp: "select",
   features: [optionsHandler],
 };
+
+declare module "@formkit/inputs" {
+  export interface FormKitInputProps<Props extends FormKitInputs<Props>> {
+    singlePageSelect: {
+      type: "singlePageSelect";
+      value?: string;
+    };
+  }
+}
